requirements.txt
python-dotenv==1.0.1
azure-cognitiveservices-speech==1.41.1
